# Conflicts:
#	enrichment/src/main/kotlin/com/rarible/protocol/union/enrichment/repository/search/EsCollectionRepository.kt
#	indexer/src/main/resources/es/migration/V1.5__collection_template.http
#	indexer/src/main/resources/mapping/collection_v0.1.json
#	pom.xml

+++ indexer/src/main/kotlin/com/rarible/protocol/union/search/indexer/handler/OwnershipEventHandler.kt
@@ -35,11 +35,15 @@ class OwnershipEventHandler(
             listOf(
                 async {
                     logger.debug("Saving ${events.size} OwnershipDto events to ElasticSearch")
-                    repository.saveAll(events)
+                    if (events.isNotEmpty()) {
+                        repository.saveAll(events)
+                    }
                 },
                 async {
                     logger.debug("Removing ${deleted.size} OwnershipDto events from ElasticSearch")
-                    repository.deleteAll(deleted)
+                    if (deleted.isNotEmpty()) {
+                        repository.deleteAll(deleted)
+                    }
                 },
             ).awaitAll()
         }
